# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: LibsClaw OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating LibsClaw HTTP APIs to /api/v1.
    Dynamic LibsClaw configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local LibsClaw server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/files:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/file:
    post:
      tags: [Open API]
      summary: Upload a file
      operationId: uploadOpenApiFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    get:
      tags: [Open API]
      summary: Download an uploaded file by attachment ID
      operationId: downloadOpenApiFile
      x-astrbot-scope: file
      parameters:
        - $ref: "#/components/parameters/AttachmentIdQuery"
      responses:
        "200":
          description: File content or an error envelope
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List skills
      operationId: listSkills
      x-astrbot-scope: skill
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Skills]
      summary: Upload or import a skill archive
      operationId: uploadSkill
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SkillUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/batch:
    post:
      tags: [Skills]
      summary: Upload multiple skill archives
      operationId: uploadSkillsBatch
      x-astrbot-scope: skill
      parameters:
        - name: on_conflict
          in: query
          description: How to handle skills that already exist locally
          schema:
            type: string
            enum: [error, skip, overwrite]
            default: error
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/inspect:
    post:
      tags: [Skills]
      summary: Inspect skill archives without installing them
      description: >-
        Returns the skills contained in each uploaded archive and whether
        they already exist locally, so clients can ask how to resolve
        duplicates before uploading.
      operationId: inspectSkills
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/by-name:
    patch:
      tags: [Skills]
      summary: Update skill metadata or enabled state by body name
      operationId: updateSkillByName
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_name]
              properties:
                skill_name:
                  type: string
                enabled:
                  type: boolean
                active:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Skills]
      summary: Delete a skill by query name
      operationId: deleteSkillByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/archive:
    get:
      tags: [Skills]
      summary: Download a skill archive by query name
      operationId: downloadSkillByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /api/v1/skills/files:
    get:
      tags: [Skills]
      summary: List files in a skill by query name
      operationId: listSkillFilesByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/file:
    get:
      tags: [Skills]
      summary: Get skill file content by query name and path
      operationId: getSkillFileByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Text"
    put:
      tags: [Skills]
      summary: Update skill file content by body name and path
      operationId: updateSkillFileByName
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_name, path, content]
              properties:
                skill_name:
                  type: string
                path:
                  type: string
                content:
                  type: string
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/{skill_name}:
    patch:
      tags: [Skills]
      summary: Update skill metadata or enabled state
      operationId: updateSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Skills]
      summary: Delete a skill
      operationId: deleteSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/{skill_name}/archive:
    get:
      tags: [Skills]
      summary: Download a skill archive
      operationId: downloadSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      responses:
        "200":
          description: Skill archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /api/v1/skills/{skill_name}/files:
    get:
      tags: [Skills]
      summary: List files in a skill
      operationId: listSkillFiles
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - name: path
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/{skill_name}/files/{file_path}:
    get:
      tags: [Skills]
      summary: Get skill file content
      operationId: getSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200":
          $ref: "#/components/responses/Text"
    put:
      tags: [Skills]
      summary: Update skill file content
      operationId: updateSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - $ref: "#/components/parameters/FilePath"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/candidates:
    get:
      tags: [Skills]
      summary: List Shipyard Neo skill candidates
      operationId: listNeoSkillCandidates
      x-astrbot-scope: skill
      parameters:
        - name: skill_key
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/releases:
    get:
      tags: [Skills]
      summary: List Shipyard Neo skill releases
      operationId: listNeoSkillReleases
      x-astrbot-scope: skill
      parameters:
        - name: skill_key
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/payload:
    get:
      tags: [Skills]
      summary: Get a Shipyard Neo skill payload
      operationId: getNeoSkillPayload
      x-astrbot-scope: skill
      parameters:
        - name: payload_ref
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/evaluate:
    post:
      tags: [Skills]
      summary: Evaluate a Shipyard Neo skill candidate
      operationId: evaluateNeoSkillCandidate
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoCandidateActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/promote:
    post:
      tags: [Skills]
      summary: Promote a Shipyard Neo skill candidate
      operationId: promoteNeoSkillCandidate
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoCandidateActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/rollback:
    post:
      tags: [Skills]
      summary: Roll back a Shipyard Neo skill release
      operationId: rollbackNeoSkillRelease
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoReleaseActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/sync:
    post:
      tags: [Skills]
      summary: Sync a Shipyard Neo skill release
      operationId: syncNeoSkillRelease
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoReleaseActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/candidates/delete:
    post:
      tags: [Skills]
      summary: Delete a Shipyard Neo skill candidate
      operationId: deleteNeoSkillCandidate
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoCandidateActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/skills/neo/releases/delete:
    post:
      tags: [Skills]
      summary: Delete a Shipyard Neo skill release
      operationId: deleteNeoSkillRelease
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeoReleaseActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard LibsClaw success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
    Text:
      description: Text response
      content:
        text/plain:
          schema:
            type: string
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    AttachmentIdQuery:
      name: attachment_id
      in: query
      required: true
      schema:
        type: string
...
    FilePath:
      name: file_path
      in: path
      required: true
      description: URL-encoded relative file path.
      schema:
        type: string
...
    SkillName:
      name: skill_name
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    FileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...
    SkillUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        overwrite:
          type: boolean
          default: false
    SkillPatchRequest:
      type: object
      properties:
        enabled:
          type: boolean
        display_name:
          type: string
        description:
          type: string
      additionalProperties: false
    NeoCandidateActionRequest:
      type: object
      required: [candidate_id]
      properties:
        candidate_id:
          type: string
      additionalProperties: true
    NeoReleaseActionRequest:
      type: object
      required: [release_id]
      properties:
        release_id:
          type: string
      additionalProperties: true
...